move verify token to the relevant file
