+++ rust/actyx/api/src/util/filters/mod.rs
@@ -2,6 +2,4 @@ mod accept;
 mod authenticate;
 
 pub(crate) use accept::{accept_json, accept_ndjson, accept_text};
-#[cfg(test)]
-pub(crate) use authenticate::verify;
 pub(crate) use authenticate::*;
